refactor: deduplicate isAnySegmentReachable across rules
