+++ lib/rules/array-callback-return.js
@@ -31,21 +31,6 @@ function isTargetMethod(node) {
 }
 
 /**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
-
-/**
  * Returns a human-legible description of an array method
  * @param {string} arrayMethodName A method name to fully qualify
  * @returns {string} the method name prefixed with `Array.` if it is a class method,
@@ -368,7 +353,7 @@ module.exports = {
 			} else {
 				if (
 					node.body.type === "BlockStatement" &&
-					isAnySegmentReachable(funcInfo.currentSegments)
+					astUtils.isAnySegmentReachable(funcInfo.currentSegments)
 				) {
 					messageAndSuggestions.messageId = funcInfo.hasReturn
 						? "expectedAtEnd"
