+++ lib/rules/getter-return.js
@@ -17,21 +17,6 @@ const astUtils = require("./utils/ast-utils");
 
 const TARGET_NODE_TYPE = /^(?:Arrow)?FunctionExpression$/u;
 
-/**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
-
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -98,7 +83,7 @@ module.exports = {
 		function checkLastSegment(node) {
 			if (
 				funcInfo.shouldCheck &&
-				isAnySegmentReachable(funcInfo.currentSegments)
+				astUtils.isAnySegmentReachable(funcInfo.currentSegments)
 			) {
 				context.report({
 					node,
