+++ lib/rules/no-fallthrough.js
@@ -9,27 +9,13 @@
 //------------------------------------------------------------------------------
 
 const { directivesPattern } = require("../shared/directives");
+const astUtils = require("./utils/ast-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
 
 const DEFAULT_FALLTHROUGH_COMMENT = /falls?\s?through/iu;
-
-/**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
 
 /**
  * Checks whether or not a given comment string is really a fallthrough comment and not an ESLint directive.
@@ -239,7 +225,7 @@ module.exports = {
 				 * `break`, `return`, or `throw` are unreachable.
 				 * And allows empty cases and the last case.
 				 */
-				const isSwitchExitReachable = isAnySegmentReachable(
+				const isSwitchExitReachable = astUtils.isAnySegmentReachable(
 					currentCodePathSegments,
 				);
 				const isFallthrough =
