+++ lib/rules/no-unreachable-loop.js
@@ -5,6 +5,8 @@
 
 "use strict";
 
+const astUtils = require("./utils/ast-utils");
+
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
@@ -16,21 +18,6 @@ const allLoopTypes = [
 	"ForInStatement",
 	"ForOfStatement",
 ];
-
-/**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
 
 /**
  * Determines whether the given node is the first node in the code path to which a loop statement
@@ -175,7 +162,7 @@ module.exports = {
 				 * Ignore unreachable loop statements to avoid unnecessary complexity in the implementation, or false positives otherwise.
 				 * For unreachable segments, the code path analysis does not raise events required for this implementation.
 				 */
-				if (isAnySegmentReachable(currentCodePathSegments)) {
+				if (astUtils.isAnySegmentReachable(currentCodePathSegments)) {
 					loopsToReport.add(node);
 				}
 			},
