+++ lib/rules/no-useless-return.js
@@ -60,21 +60,6 @@ function isInFinally(node) {
 	return false;
 }
 
-/**
- * Checks all segments in a set and returns true if any are reachable.
- * @param {Set<CodePathSegment>} segments The segments to check.
- * @returns {boolean} True if any segment is reachable; false otherwise.
- */
-function isAnySegmentReachable(segments) {
-	for (const segment of segments) {
-		if (segment.reachable) {
-			return true;
-		}
-	}
-
-	return false;
-}
-
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -346,7 +331,7 @@ module.exports = {
 					astUtils.isInLoop(node) ||
 					isInFinally(node) ||
 					// Ignore `return` statements in unreachable places (https://github.com/eslint/eslint/issues/11647).
-					!isAnySegmentReachable(scopeInfo.currentSegments)
+					!astUtils.isAnySegmentReachable(scopeInfo.currentSegments)
 				) {
 					return;
 				}
